+++ cmd/tailscaled/tailscaled_windows.go
@@ -30,6 +30,7 @@ import (
 	"golang.zx2c4.com/wireguard/windows/tunnel/winipcfg"
 	"tailscale.com/ipn/ipnserver"
 	"tailscale.com/logpolicy"
+	"tailscale.com/net/tun"
 	"tailscale.com/tempfork/wireguard-windows/firewall"
 	"tailscale.com/types/logger"
 	"tailscale.com/version"
@@ -159,11 +160,16 @@ func startIPNServer(ctx context.Context, logid string) error {
 	var err error
 
 	getEngine := func() (wgengine.Engine, error) {
+		dev, err := tun.New(logf, "Tailscale")
+		if err != nil {
+			return nil, err
+		}
 		eng, err := wgengine.NewUserspaceEngine(logf, wgengine.Config{
-			TUNName:    "Tailscale",
+			TUN:        dev,
 			ListenPort: 41641,
 		})
 		if err != nil {
+			dev.Close()
 			return nil, err
 		}
 		return wgengine.NewWatchdog(eng), nil
